added documentation to the files
