--- boilerplate/src/configs/classNames.js
@@ -1,15 +1,0 @@
-const names = {
-  /*
-    Models
-  */
-
-  demo: 'demo',
-
-  /*
-     Repository
-  */
-
-  knexManager: 'knex-manager',
-};
-
-module.exports = names;
