+++ boilerplate/src/configs/modelNames.js
@@ -1,0 +1,21 @@
+/**
+ * This class defines global constants for model-names.
+ * So that any change in file name will require only changing them here
+ * key:value, where value is the actual model name
+ */
+
+const names = {
+  /*
+    Models
+  */
+
+  demo: 'demo',
+
+  /*
+     Repository
+  */
+
+  knexManager: 'knex-manager',
+};
+
+module.exports = names;
