+++ boilerplate/src/endpoints/demo/handlers.js
@@ -1,11 +1,16 @@
+/**
+ * This is a handler for all the requests to `demo` resource.
+ * It extends the base handler class to leverage all the default routes.
+ */
+
 'use strict';
 
 const elv = require('elv');
-const BaseHandlers = require('../base-handlers');
+const BaseHandler = require('../base-handler');
 
 const Model = require('../../models').demo.Model;
 
-class Handlers extends BaseHandlers {
+class Handlers extends BaseHandler {
   constructor(model) {
     const m = elv.coalesce(model, () => new Model());
     super(m);
