+++ boilerplate/src/models/demo/index.js
@@ -1,3 +1,7 @@
+/**
+ * Exports the model class
+ */
+
 'use strict';
 
 const Model = require('./model');
