+++ boilerplate/src/models/demo/repository.js
@@ -1,16 +1,21 @@
+/**
+ * This class defines all the methods to handle calls to db for `demo` resource,
+ * using query-builder tool
+ */
+
 'use strict';
 
 const PostgreSqlRepository = require('../../repository');
 
 const logger = require('../../utils/logger');
-const Names = require('../../configs/classNames');
+const Names = require('../../configs/modelNames');
 const Kernel = require('../kernel');
 
 const resource = Names.demo;
 
 class DemoRepository extends PostgreSqlRepository {
   constructor() {
-    const knexManager = Kernel.resolve(Names.knexManager);
+    const knexManager = Kernel.resolve(Names.knexManager); // pick the already defined knex-instance
 
     super({
       name: Names.demo,
