+++ boilerplate/src/models/index.js
@@ -1,9 +1,12 @@
+/**
+ * This class exports all the models defined in the models folder
+ */
+
 'use strict';
 
 const Kernel = require('./kernel');
 
 // DEMO
-//
 const demo = require('./demo');
 
 module.exports = {
