+++ boilerplate/src/models/knex-manager.js
@@ -1,9 +1,10 @@
-const Kernel = require('./kernel');
 /**
-* Kernel is used to better manage Knex instances across the app.
-
-Because Kernel bindings must be objects, we created a Knex manager class.
+ * Kernel is used to better manage Knex instances across the app.
+ Because Kernel bindings must be objects, we created a Knex manager class.
  */
+
+const Kernel = require('./kernel');
+
 class KnexManager {
   constructor() {
     this._knex = null;
